+++ CHANGELOG.md
@@ -9,6 +9,8 @@
  - `ImageViewer.getImageMetadata()` returns the imagedata for the image being viewed.
  - `ImageViewer.getImageBounds()` returns the current bounds of the viewport in image pixels (note: _not_ screen pixels, image pixels are pixels in relation to the intrinsic image size, not it's size on the screen).
  - `ImageViewer.fitImageBounds(bounds)` fits the viewer to `bounds`.
+ - `ImageViewerMarker` class added. `ImageViewer`can now have markers positioned in image pixels.
+
 
 ### 🐛 Bug fixes
 - fixes a bug where `map.getProjection()` did not return a value when default projection was used